Allow manually setting pattern for testing (#1910)
## Description:

This pr allows setting the pattern manually in using the dev console so
we can see how it looks before uploading.

To set it: set the b64 pattern in local storage with key: dev-pattern.

This will override set pattern. Only works in singleplayer mode.

## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced

## Please put your Discord username so you can be contacted if a bug or
regression is found:

evan