Keep Lobby ID bar sticky at top of modal (#2108)
## Description:

This PR adjusts the HostLobbyModal so that the lobby ID bar stays
visible at the top of the modal when scrolling.
Fixes #2082 


https://github.com/user-attachments/assets/273b6ebe-3538-4489-b771-67261f7642f8


In this PR, when the ID box became sticky, the background sometimes made
the ID hard to read.
To fix this, I made the background color slightly darker for better
visibility.

## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced

## Please put your Discord username so you can be contacted if a bug or
regression is found:

aotumuri