remove unused map-generator/generated directory (#2107)
## Description:

This PR removes the `map-generator/generated` directory to avoid
confusion.


The current map generation process (`map-generator/main.go`) outputs map
data to:
  - `../resources/maps` (for normal maps)
  - `../tests/testdata/maps` (for test maps)
The `generated` directory is not used anywhere in the current codebase.

Keeping the unused `map-generator/generated` directory may cause
confusion for contributors or users who expect it to be part of the
active map pipeline. Removing it clarifies that only `resources/maps`
and `testdata/maps` are valid output directories.

No functional changes are introduced.
## Please complete the following:

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced

## Please put your Discord username so you can be contacted if a bug or
regression is found:

aotumuri